requirements.txt
contourpy==1.3.3
cycler==0.12.1
flake8==7.3.0
fonttools==4.62.1
kiwisolver==1.5.0
librt==0.9.0
matplotlib==3.10.8
mccabe==0.7.0
mypy==1.20.2
mypy_extensions==1.1.0
numpy==2.4.4
packaging==26.1
pathspec==1.1.0
pillow==12.2.0
pycodestyle==2.14.0
pyflakes==3.4.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.15.0
